requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.2
cloudinary==1.44.1
constantly==23.10.4
cryptography==46.0.6
daphne==4.2.1
defusedxml==0.7.1
Django==5.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-mongodb-backend==5.2.0b1
django-pwa==2.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.3
dnspython==2.7.0
drf-spectacular==0.28.0
geographiclib==2.1
geopy==2.4.1
google-api-core==2.25.2
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
gunicorn==25.1.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
msgpack==1.1.2
oauthlib==3.3.1
packaging==26.0
phonenumbers==9.0.12
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pymongo==4.13.2
pyOpenSSL==26.0.0
pyparsing==3.2.3
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
PyYAML==6.0.2
querystring-parser==1.2.4
redis==7.4.0
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
rsa==4.9.1
service-identity==24.2.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tomli==2.4.1
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
zope.interface==8.2
